extractor/ppt_extractor.py
# extractor/ppt_extractor.py
from pptx import Presentation

def extract_text_pptx(path):
    prs = Presentation(path)
    texts = []
    for slide in prs.slides:
        for shape in slide.shapes:
            if hasattr(shape, "text") and shape.text.strip():
                texts.append(shape.text)
    return "\n".join(texts)
