requirements.txt
selenium
webdriver-manager
pymongo
requests
python-dotenv
beautifulsoup4
lxml
python-pptx
pymupdf
pillow
pytesseract
easyocr
tqdm
